+++ docs/changelog.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2026-01-15
+
+### Added
+
+ - Add config options `tcp-retry-count`, `tcp-retry-redispatch`, `tcp-load-balancing-algorithm`, 
+  `tcp-load-balancing-consistent-hashing`.
+ - Fix bug where integrating with `haproxy-route-tcp` does not update the relation data.
+
 ## 2026-01-12
 
 ### Added